refactor: moved AdminController methods
